% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logMarPre.R
\name{logMarPre}
\alias{logMarPre}
\title{Log-Marginal Distribution of a Normal with Unknown Presicion}
\usage{
logMarPre(y, D, b)
}
\arguments{
\item{y}{A list containing the observations of the multivariate normal.}

\item{D}{Prior scale matrix for the Wishart Distribution, note that this is
not inverted inside the trace.}

\item{b}{Total degrees of freedom. Again, this differs to the traditional
Wishart representation.}
}
\value{
The value of the log-marginal distribution.
}
\description{
This function computes the log-marginal distribution of multivariate normal
data with zero mean and unknown variance.
}
\author{
Rene Gutierrez Marquez
}
