% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMBTC.R
\name{PMBTC}
\alias{PMBTC}
\title{Pseudo Marginal Bayesian Tensor Clustering for Precision Matrices}
\usage{
PMBTC(iP, theta, kPseCov, burnin = 0, nmcmc = 100, progress = TRUE)
}
\arguments{
\item{iP}{A list specifying the Initial Partition.}

\item{theta}{Dirichlet Process Prior Parameter.}

\item{kPseCov}{List of Vecorized k-Matrization Pseudo Covariances for each
dimension and for each observation.}

\item{burnin}{Number of sample to burn in in the MCMC.}

\item{nmcmc}{Number of MCMC samples desired as output. That is without
considering the burn-in period.}

\item{progress}{Boolean indicating the use of the progress bar. By default
it is set true.}
}
\value{
A list containing samples from the posterior partition.
}
\description{
This function performs clustering of Tensor Normal samples with zero mean
and Different Precision Matrices throught the use of Pseudo Data Marginal
distributions. Based on the Pseudo-Marginal the procedure performs Chinese
Restaurant Stochastic Search based on a Dirichlet Process Prior.
}
\author{
Rene Gutierrez Marquez
}
