% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proMat.R
\name{proMat}
\alias{proMat}
\title{Observations Probability Matrix}
\usage{
proMat(sP, n)
}
\arguments{
\item{sP}{A list containing posterior sample observations.}

\item{n}{NUmber of observations.}
}
\value{
The Probability Matrix.
}
\description{
This function computes the probability matrix following Lau and Green (2007)
  according to a posterior sample Partition. The posterior probability
  matrix that specifies the posterior probablility that 2 observations
  belong to the same cluster.
}
\author{
Rene Gutierrez Marquez
}
