% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenCovPar.R
\name{tenCovPar}
\alias{tenCovPar}
\title{Tensor Data Splitter by Partition}
\usage{
tenCovPar(P, pseCov)
}
\arguments{
\item{P}{A list containing a partition of indexes.}

\item{pseCov}{}
}
\value{
A list of lists containing the ...
}
\description{
Partitions
}
\author{
Rene Gutierrez Marquez
}
