% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenPrePseMar.R
\name{tenPrePseMar}
\alias{tenPrePseMar}
\title{Pseudo Marginal Distribution of Normal Tensor}
\usage{
tenPrePseMar(C)
}
\arguments{
\item{C}{A list of matrices, one for each dimension of the tensor containing
the matricization of the tensor by each dimension.}
}
\value{
The value of the pseudo log-marginal distribution.
}
\description{
Evaluates the a Pseudo Marginal Distribution of Normal Tensor Data with zero
mean and precision matrices are unknown. Note that this is not the actual
marginal or proportional to it.
}
\author{
Rene Gutierrez Marquez
}
